Add categorys rest controller + view

+++ src/main/java/net/ivan/kavaliou/moneyman/service/TransactionCategoryService.java
@@ -1,7 +1,10 @@
 package net.ivan.kavaliou.moneyman.service;
 
+import net.ivan.kavaliou.moneyman.exceptions.NotFoundException;
+import net.ivan.kavaliou.moneyman.exceptions.ServiceException;
 import net.ivan.kavaliou.moneyman.model.persistence.TransactionCategory;
 import net.ivan.kavaliou.moneyman.repository.TransactionCategoryRepository;
+import net.ivan.kavaliou.moneyman.utils.Messages;
 import net.ivan.kavaliou.moneyman.utils.enums.TransactionType;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
@@ -21,6 +24,12 @@ public class TransactionCategoryService {
     @Autowired
     TransactionTypesService transactionTypesService;
 
+    @Autowired
+    TransactionService transactionService;
+
+    @Autowired
+    Messages messages;
+
     public Optional<TransactionCategory> get(Integer id){
         return repository.findByUserAndId(usersService.getAuthUser(), id);
     }
@@ -28,10 +37,13 @@ public class TransactionCategoryService {
     public boolean delete(Integer id){
         Optional<TransactionCategory> category = get(id);
         if (category.isPresent()){
-            repository.delete(category.get());
-            return true;
+            if (transactionService.getAllByCategory(category.get()).size() == 0){
+                repository.delete(category.get());
+                return true;
+            }
+            throw new ServiceException(messages.get("error.category.transactions.exsist"));
         }
-        return false;
+        throw new NotFoundException(messages.get("error.category.notFound"));
     }
 
     public List<TransactionCategory> getExpenses(){
@@ -40,6 +52,13 @@ public class TransactionCategoryService {
 
     public List<TransactionCategory> getIncomes(){
         return repository.findByUserAndAndTransactionType(usersService.getAuthUser(), transactionTypesService.getIncome());
+    }
+
+    public TransactionCategory add(TransactionCategory transactionCategory){
+        if (repository.findOneByUserAndAndTransactionTypeAndName(usersService.getAuthUser(),transactionCategory.getTransactionType(),transactionCategory.getName()).isPresent()){
+            throw new ServiceException("error.category.exsist");
+        }
+        return save(transactionCategory);
     }
 
     public TransactionCategory save(TransactionCategory transactionCategory){
